Fixing ordering of files on upload step.
Moved group_files to upload controller.

Fixes #83
